[Bug 409]:   Integer arguments passed as void pointers are cast to long.
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@15899 9d14d920-75d5-0310-8bd7-a52572981aee

+++ agents/linux/ftp_routines.c
@@ -444,7 +444,7 @@ read_test(void *arg)
         if (ret < 0)
         {
             ERROR("FTP get select() error errno=%d", errno);
-            return (void *)errno;
+            return (void *)(long)errno;
         }
         if (ret == 0)
         {
@@ -458,7 +458,7 @@ read_test(void *arg)
             if (l < 0)
             {
                 ERROR("FTP get read() error errno=%d", errno);
-                return (void *)errno;
+                return (void *)(long)errno;
             }
             if (l == 0)
             {
@@ -510,7 +510,7 @@ write_test(void *arg_void)
             if (errno != EAGAIN)
             {
                 ERROR("FTP put write() error errno=%d", errno);
-                return (void *)errno;
+                return (void *)(long)errno;
             }
 
             gettimeofday(&now, NULL);
@@ -532,7 +532,7 @@ write_test(void *arg_void)
                 if (ret < 0)
                 {
                     ERROR("select() errno=%d", errno);
-                    return (void *)errno;
+                    return (void *)(long)errno;
                 }
             }
             else
@@ -596,7 +596,8 @@ ftp_test(char *uri_get, char *uri_put, int size)
     
     VERB("Open OK\n");
 
-    if (si >= 0 && pthread_create(&ti, NULL, read_test, (void *)si) < 0)
+    if (si >= 0 && pthread_create(&ti, NULL, read_test, 
+                                  (void *)(long)si) < 0)
     {
         int rc = errno;
